Use PromiEvent subscriptions to listen for Tx results
Web3.js 0.20 had eth.filter
  - https://github.com/ChainSafe/web3.js/blob/0.20.7/DOCUMENTATION.md#web3ethfilter

But it doesn't exist anymore in Web3 1.7.

Instead more straight-forward way is to use the PromiEvent subscriptions
  - https://web3js.readthedocs.io/en/v1.7.1/callbacks-promises-events.html

This simplifies the code quite a bit.

Still - the `start-listener!` functionality is broken and `:web3/watch-transactions`
implementation was replaced with one that just throws error (it's used
in 2 places in `hegex/src/district_hegex/ui/trading/events.cljs`)
  - https://github.com/district0x/hegex/blob/62783ea30e8ebd997945d8e62dc2261233415293/src/district_hegex/ui/trading/events.cljs#L345-L346